Fixed typos "childchild"
This is a fantastic library, thank you so much for making it available. I just discovered that the `contains` method was not working properly and after investigating found that it and a few others called node.rightchildchild instead of node.rightchild (or the leftchild versions of those). This s a quick fix for that.